+++ src/task_allocation/src/partitionable_polled_task.cpp
@@ -40,8 +40,9 @@ partitionable_polled_task::partitionable_polled_task(
       partitionable_task(c_params) {}
 
 void partitionable_polled_task::init_random(const polled_task *const partition,
-                                            int lb, int ub) {
-  executable_task::init_exec_estimate(std::rand() % (ub - lb + 1) + lb);
+                                            uint lb, uint ub) {
+  executable_task::init_exec_estimate(
+      static_cast<uint>(std::rand()) % (ub - lb + 1) + lb);
   last_partition(partition);
 } /* init_random() */
